feat: project structure merged with dockerization Ok
