+++ frontend/Dockerfile
@@ -10,7 +10,8 @@ RUN apt-get update \
     && apt-get install -y --no-install-recommends build-essential \
     && rm -rf /var/lib/apt/lists/*
 
-COPY requirements.txt .
+# Copiar frontend requirements e instalar dependencias Python
+COPY frontend/requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
 RUN useradd --create-home appuser
