--- requirements.txt
@@ -1,29 +1,0 @@
-# Model Explainability
-shap==0.43.0
-
-# Utils
-python-dotenv==1.0.0
-pyyaml==6.0.1
-
-# Testing
-pytest==7.4.3
-pytest-cov==4.1.0
-
-numpy==2.3.3
-pandas==2.3.3
-scikit-learn==1.7.2
-scipy==1.16.2
-imbalanced-learn==0.11.0
-joblib==1.5.2
-jupyter==1.0.0
-notebook==7.0.6
-ipykernel==6.30.1
-streamlit==1.38.0
-matplotlib==3.9.2
-seaborn==0.13.2
-plotly==5.24.1
-fastapi>=0.111.0
-uvicorn[standard]>=0.29.0
-pydantic>=2.9.2
-python-multipart==0.0.9
-requests==2.31.0
